requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
beautifulsoup4==4.10.0; python_version >= '3.1'
boto3==1.18.42
botocore==1.21.42; python_version >= '3.6'
cerberus==1.3.4
certifi==2021.5.30
charset-normalizer==2.0.5; python_version >= '3.0'
click==8.0.1
cloudflare==2.8.15
deepdiff==5.5.0
idna==3.2; python_version >= '3.0'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonlines==2.0.0; python_version >= '3.6'
loguru==0.5.3
ordered-set==4.0.2; python_version >= '3.5'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-magic==0.4.24
python-slugify==5.0.2
pyyaml==5.4.1
requests==2.26.0
s3transfer==0.5.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.2.1; python_version >= '3.6'
text-unidecode==1.3
tqdm==4.62.2
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
